+++ packages/components/inputs/multiline-text-input/src/index.js
@@ -1,1 +1,3 @@
 export { default } from './multiline-text-input';
+
+export { default as version } from './version';
