+++ packages/components/inputs/multiline-text-input/src/multiline-text-input.story.js
@@ -3,8 +3,8 @@ import { Value } from 'react-value';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean, text, select } from '@storybook/addon-knobs/react';
-import Section from '../../../../.storybook/decorators/section';
-import Readme from './README.md';
+import Section from '../../../../../.storybook/decorators/section';
+import Readme from '../README.md';
 import MultilineTextInput from './multiline-text-input';
 
 storiesOf('Components|Inputs', module)
